Move Database to base file for discovery
