+++ tests/functional/demo/test_demo.py
@@ -8,7 +8,7 @@ import numpy
 import pytest
 
 import orion.core.cli
-from orion.core.io.database import Database
+from orion.core.io.database.base import Database
 from orion.core.worker import workon
 from orion.core.worker.experiment import Experiment
