+++ tests/unittests/core/conftest.py
@@ -11,7 +11,7 @@ import pytest
 from orion.algo.space import (Categorical, Integer, Real, Space)
 from orion.core.evc import conflicts
 from orion.core.io.convert import (JSONConverter, YAMLConverter)
-from orion.core.io.database import Database
+from orion.core.io.database.base import Database
 from orion.core.io.space_builder import DimensionBuilder
 from orion.core.worker.experiment import Experiment
 
@@ -124,7 +124,7 @@ def hacked_exp(with_user_dendi, random_dt, clean_db, create_db_instance):
     fake database.
     """
     try:
-        Database(of_type=('orion.core.io.database.mongodb', 'MongoDB'), name='orion_test',
+        Database('MongoDB', name='orion_test',
                  username='user', password='pass')
     except (TypeError, ValueError):
         pass
